Move conflict check to config_flow #221
